fix: remove slip_in_column_break()
This function is about to be deprecated, this PR make your package valid for CRAN. It can be removed because it does not have effect as it is positioned at the end of the section. However, there is no need for it as Word is doing the expected 2 columns rendering.